pyproject.toml
[tool.poetry]
name ="experience_trainer"
version = "0.1.0"
description = ""
authors = ["frapercan <frapercan1@alum.us.es>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.1"
pytorch-lightning = "^2.0.3"
torchvision = "^0.15.2"
webdataset = "^0.2.48"
matplotlib = "^3.7.1"
pillow = "^9.5.0"
pyyaml = "^6.0"
transformers = "^4.32.1"
diffusers = "^0.20.2"
accelerate = "^0.22.0"
timm = "^0.9.7"
torchviz = "^0.0.2"


[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
coverage = {extras = ["toml"], version = "^7.2.1"}
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.10.3"

[tool.coverage.run]
omit = ["*/site-packages/*"]

[tool.coverage.report]
fail_under = 100

[tool.taskipy.tasks]
html_docs = "make html -C docs"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

